// packages/forms/frontend/sirius-components-forms/src/propertysections/imagepicker/ImagePickerTabs.types.ts
/*******************************************************************************
 * Copyright (c) 2024 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/
import { ImageFilterCategory } from './ImageCategoryFilter.types';
import { GQLCustomImage } from './ImagePickerPropertySection.types';

export interface ImagePickerGalleryTabProps {
  onSelection: (image: GQLCustomImage) => void;
  images: GQLCustomImage[];
  selection: GQLCustomImage | null;
}

export interface ImagePickerLinkTabProps {
  // TODO add props of the Link tab here
}

export interface ImagePickerGalleryTabState {
  filterBarText: string;
  filteredImages: GQLCustomImage[];
  filterCategory: ImageFilterCategory;
}